+++ src/ecryptfs_fmt_plug.c
@@ -60,7 +60,7 @@ static int omp_t = 1;
 #define MIN_KEYS_PER_CRYPT		1
 #ifdef SIMD_COEF_64
 #define MAX_KEYS_PER_CRYPT      SIMD_COEF_64
-#define GETPOS_512(i, index)    ( (index&(SIMD_COEF_64-1))*8 + ((i)&(0xffffffff-7))*SIMD_COEF_64 + (7-((i)&7)) + index/SIMD_COEF_64*SHA512_BUF_SIZ*SIMD_COEF_64 *8 )
+#define GETPOS_512(i, index)    ( (index&(SIMD_COEF_64-1))*8 + ((i)&(0xffffffff-7))*SIMD_COEF_64 + (7-((i)&7)) + (unsigned int)index/SIMD_COEF_64*SHA512_BUF_SIZ*SIMD_COEF_64 *8 )
 #else
 #define MAX_KEYS_PER_CRYPT		1
 #endif
@@ -210,7 +210,7 @@ static int crypt_all(int *pcount, struct db_salt *salt)
 		SHA512_CTX ctx;
 #ifdef SIMD_COEF_64
 		unsigned char tmpBuf[64];
-		int i;
+		unsigned int i;
 		unsigned char _IBuf[128*SIMD_COEF_64+MEM_ALIGN_SIMD], *keys;
 		ARCH_WORD_64 *keys64;
