+++ src/keyring_fmt_plug.c
@@ -51,7 +51,7 @@ john_register_one(&fmt_keyring);
 #define MIN_KEYS_PER_CRYPT	1
 #ifdef SIMD_COEF_32
 #define MAX_KEYS_PER_CRYPT SIMD_COEF_32
-#define GETPOS(i, index)        ( (index&(SIMD_COEF_32-1))*4 + ((i)&(0xffffffff-3))*SIMD_COEF_32 + (3-((i)&3)) + index/SIMD_COEF_32*SHA256_BUF_SIZ*SIMD_COEF_32*4 )
+#define GETPOS(i, index)        ( (index&(SIMD_COEF_32-1))*4 + ((i)&(0xffffffff-3))*SIMD_COEF_32 + (3-((i)&3)) + (unsigned int)index/SIMD_COEF_32*SHA256_BUF_SIZ*SIMD_COEF_32*4 )
 #else
 #define MAX_KEYS_PER_CRYPT	1
 #endif
@@ -209,7 +209,7 @@ static void symkey_generate_simple(int index, unsigned char *salt, int n_salt, i
 	SHA256_CTX ctx;
 	unsigned char digest[32], _IBuf[64*SIMD_COEF_32+MEM_ALIGN_SIMD], *keys;
 	uint32_t *keys32;
-	int i, j;
+	unsigned int i, j;
 
 	keys = (unsigned char*)mem_align(_IBuf, MEM_ALIGN_SIMD);
 	memset(keys, 0, 64*SIMD_COEF_32);
